+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v6.183](https://github.com/replicant4j/replicant/tree/v6.183) (2026-03-30) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.182...v6.183)
+
+Changes in this release:
 
 * ELiminate the need for `ReplicantContextHolder.remove()`.
 * Make `ReplicantContextHolder.clear()` public.